// packages/pulse/src/bundler/types.ts
export interface PulseConfig {
  root: string;
  srcDir: string;
  outDir: string;
  publicDir: string;
  pages: {
    dir: string;
    defaultLayout?: string;
  };
  build: {
    minify: boolean;
    sourcemap: boolean;
    target: 'es2020' | 'es2022' | 'esnext';
    splitting: boolean;
    treeshake: boolean;
    islands: boolean;
    ssr: boolean;
  };
  optimization: {
    inlineStyles: boolean;
    criticalCSS: boolean;
    prefetch: boolean;
    compress: 'gzip' | 'brotli' | 'both';
  };
  devServer: {
    port: number;
    hmr: boolean;
    open: boolean;
  };
  debug: boolean;
}